게시판 1차 버전.

+++ test/src/main/java/test/dto/fileDTO.java
@@ -8,15 +8,15 @@ import lombok.Data;
 public class fileDTO {
 	
 	// 파일 테이블
-	private int file_id;
-	private int user_no;
-	private int post_id;
+	private Long file_id;
+	private Long user_no;
+	private Long post_id;
 	private String file_name;
 	private String file_path;
-	private int file_size;
+	private Long file_size;
 	private String file_type;
-	private int width;
-	private int heigth;
+	private Integer width;
+	private Integer heigth;
 	private String thumnail_path; 
 	private Date reg_date;
 	private Date udt_date;
